+++ package.json
@@ -14,6 +14,7 @@
     "@hypercerts-org/sdk": "^2.2.0",
     "@radix-ui/react-accordion": "^1.2.1",
     "@radix-ui/react-alert-dialog": "^1.1.2",
+    "@radix-ui/react-progress": "^1.1.0",
     "@radix-ui/react-select": "^2.1.2",
     "@remix-run/node": "^2.13.1",
     "@remix-run/react": "^2.13.1",
@@ -31,6 +32,7 @@
     "lucide-react": "^0.454.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.53.1",
     "seedrandom": "^3.0.5",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
